Attempt fixing the map.
THE ISSUE: It appears that matchArr() and objectives() computed
properties are just never being loaded because nothing constantly
depends on them to force them to update. Putting a const referencing
them in mounted does not work because it will only load once BEFORE the
api data is even loaded. Need to move map to another Vue instance and
load it into the page using props probably…